requirements.txt
streamlit
pandas
sqlalchemy
pymysql
psycopg2-binary
st-pages
streamlit-extras
yt-dlp>=2023.10.13